refactor: hooks 밖으로 빼기
